moved test classes to correct package
